Add logger module
Implementing separated logger module.
