+++ AudioEncode/CMakeLists.txt
@@ -11,7 +11,7 @@ set(CPPS
 add_library(AudioEncode STATIC ${HEADERS} ${CPPS})
 target_link_libraries(AudioEncode PRIVATE
   libacm
-  plog::plog
+  logger
 )
 target_include_directories(AudioEncode PUBLIC
   $<BUILD_INTERFACE:
