+++ CMakeLists.txt
@@ -234,6 +234,7 @@ add_subdirectory(linux)
 add_subdirectory(ddio)
 add_subdirectory(dd_video)
 add_subdirectory(fix)
+add_subdirectory(logger)
 add_subdirectory(manage)
 add_subdirectory(grtext)
 add_subdirectory(mem)
