+++ dd_grwin32/CMakeLists.txt
@@ -11,8 +11,8 @@ set(CPPS
 add_library(dd_grwin32 STATIC ${HEADERS} ${CPPS})
 target_link_libraries(dd_grwin32 PRIVATE
   2dlib
+  logger
   misc
-  plog::plog
 )
 target_compile_definitions(dd_grwin32 PRIVATE DX_APP)
 target_link_libraries(dd_grwin32 PRIVATE ${DDRAW_LIBRARY})
