+++ ddio/CMakeLists.txt
@@ -25,8 +25,8 @@ set(CPPS
 add_library(ddio STATIC ${HEADERS} ${CPPS})
 target_link_libraries(ddio PRIVATE
   SDL2::SDL2
-  plog::plog
   ddebug
+  logger
   mem
   misc
 )
