+++ manage/CMakeLists.txt
@@ -25,10 +25,10 @@ add_library(manage STATIC ${HEADERS} ${CPPS})
 target_link_libraries(manage PRIVATE
   cfile
   ddio
+  logger
   mem
   misc
   model
   sndlib
   stream_audio
-  plog::plog
 )
