+++ module/CMakeLists.txt
@@ -8,6 +8,7 @@ set(CPPS
 add_library(module STATIC ${HEADERS} ${CPPS})
 target_link_libraries(module PRIVATE
   ddebug
+  logger
 )
 target_include_directories(module PUBLIC
   $<BUILD_INTERFACE:
