+++ networking/CMakeLists.txt
@@ -13,9 +13,9 @@ endif()
 add_library(networking STATIC ${HEADERS} ${CPPS})
 target_link_libraries(networking PRIVATE
   ddio
+  logger
   mem
   misc
   module
   ${PLATFORM_LIBS}
-  plog::plog
 )
